refactor filter query
