+++ x-pack/plugins/lens/public/shared_components/filter_query_input.tsx
@@ -1,0 +1,144 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+import React, { useCallback, useState } from 'react';
+import { i18n } from '@kbn/i18n';
+import {
+  EuiLink,
+  EuiPanel,
+  EuiPopover,
+  EuiFormRow,
+  EuiFlexItem,
+  EuiFlexGroup,
+  EuiIconTip,
+  EuiPopoverProps,
+} from '@elastic/eui';
+import type { Query } from '@kbn/es-query';
+import { QueryInput, useDebouncedValue, validateQuery } from '.';
+import type { IndexPattern } from '../types';
+
+const filterByLabel = i18n.translate('xpack.lens.indexPattern.filterBy.label', {
+  defaultMessage: 'Filter by',
+});
+
+// to do: get the language from uiSettings
+export const defaultFilter: Query = {
+  query: '',
+  language: 'kuery',
+};
+
+export function FilterQueryInput({
+  inputFilter,
+  onChange,
+  indexPattern,
+  helpMessage,
+  label = filterByLabel,
+}: {
+  inputFilter: Query | undefined;
+  onChange: (query?: Query) => void;
+  indexPattern: IndexPattern;
+  helpMessage: string | null;
+  label?: string;
+}) {
+  const [filterPopoverOpen, setFilterPopoverOpen] = useState(false);
+  const { inputValue: queryInput, handleInputChange: setQueryInput } = useDebouncedValue<Query>({
+    value: inputFilter ?? defaultFilter,
+    onChange,
+  });
+
+  const onClosePopup: EuiPopoverProps['closePopover'] = useCallback(() => {
+    setFilterPopoverOpen(false);
+    if (inputFilter) {
+      setQueryInput(inputFilter);
+    }
+  }, [inputFilter, setQueryInput]);
+
+  const { isValid: isInputFilterValid } = validateQuery(inputFilter, indexPattern);
+  const { isValid: isQueryInputValid, error: queryInputError } = validateQuery(
+    queryInput,
+    indexPattern
+  );
+
+  return (
+    <EuiFormRow
+      display="rowCompressed"
+      label={
+        helpMessage ? (
+          <>
+            {label}{' '}
+            <EuiIconTip
+              color="subdued"
+              content={helpMessage}
+              iconProps={{
+                className: 'eui-alignTop',
+              }}
+              position="top"
+              size="s"
+              type="questionInCircle"
+            />
+          </>
+        ) : (
+          label
+        )
+      }
+      fullWidth
+      isInvalid={!isInputFilterValid}
+    >
+      <EuiFlexGroup gutterSize="s" alignItems="center">
+        <EuiFlexItem>
+          <EuiPopover
+            isOpen={filterPopoverOpen}
+            closePopover={onClosePopup}
+            anchorClassName="eui-fullWidth"
+            panelClassName="lnsIndexPatternDimensionEditor__filtersEditor"
+            button={
+              <EuiPanel paddingSize="none" hasShadow={false} hasBorder>
+                <EuiFlexGroup gutterSize="s" alignItems="center" responsive={false}>
+                  <EuiFlexItem grow={false}>{/* Empty for spacing */}</EuiFlexItem>
+                  <EuiFlexItem grow={true}>
+                    <EuiLink
+                      className="lnsFiltersOperation__popoverButton"
+                      data-test-subj="indexPattern-filters-existingFilterTrigger"
+                      onClick={() => {
+                        setFilterPopoverOpen(!filterPopoverOpen);
+                      }}
+                      color={isInputFilterValid ? 'text' : 'danger'}
+                      title={i18n.translate('xpack.lens.indexPattern.filterBy.clickToEdit', {
+                        defaultMessage: 'Click to edit',
+                      })}
+                    >
+                      {inputFilter?.query ||
+                        i18n.translate('xpack.lens.indexPattern.filterBy.emptyFilterQuery', {
+                          defaultMessage: '(empty)',
+                        })}
+                    </EuiLink>
+                  </EuiFlexItem>
+                </EuiFlexGroup>
+              </EuiPanel>
+            }
+          >
+            <EuiFormRow
+              label={label}
+              isInvalid={!isQueryInputValid}
+              error={queryInputError}
+              fullWidth={true}
+              data-test-subj="indexPattern-filter-by-input"
+            >
+              <QueryInput
+                indexPatternTitle={indexPattern.title}
+                disableAutoFocus={true}
+                value={queryInput}
+                onChange={setQueryInput}
+                isInvalid={!isQueryInputValid}
+                onSubmit={() => {}}
+              />
+            </EuiFormRow>
+          </EuiPopover>
+        </EuiFlexItem>
+      </EuiFlexGroup>
+    </EuiFormRow>
+  );
+}
