+++ x-pack/plugins/lens/public/shared_components/index.ts
@@ -36,5 +36,6 @@ export { NameInput } from './name_input';
 export { ValueLabelsSettings } from './value_labels_settings';
 export { AxisTitleSettings } from './axis_title_settings';
 export { DimensionEditorSection } from './dimension_section';
+export { FilterQueryInput } from './filter_query_input';
 export * from './static_header';
 export * from './vis_label';
